% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/001_find_base_visit.R
\name{make_drug_baseline_visit_dataset}
\alias{make_drug_baseline_visit_dataset}
\title{Title}
\usage{
make_drug_baseline_visit_dataset(
  visits_df,
  drug_df,
  target_generic_key,
  baseline_cutoff_days = 183,
  id_col = id,
  visitdate_col = visitdate,
  indexn_col = indexn,
  indexN_col = indexN,
  generic_key_col = generic_key,
  init_generic_col = init_generic,
  generic_start_date_col = generic_start_date,
  assign_output = FALSE,
  output_env = .GlobalEnv
)
}
\arguments{
\item{output_env}{}
}
\description{
Title
}
\examples{

# Step 0, find visits and drug data
\dontrun{
visits <- readRDS("visits_data.rds")
drug <- readRDS("drugexpdetails_data.rds")

# Step 1, find baseline visit with user defined cutoff days. Currently using 183 days.

init_upadacitinib<- corallinits::make_drug_baseline_visit_dataset(
  visits_df = visits,
  drug_df = drug,
  target_generic_key = "upadacitinib",
  baseline_cutoff_days = 183
)
}
}
